backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
celery==5.3.6
redis==5.0.1
boto3==1.34.30
python-dateutil==2.8.2
pytest==8.3.3
httpx==0.27.2
pytest-asyncio==0.24.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
bcrypt==4.0.1
